Printing out a phrase using string concatenation
I assigned each word to one of the variables provided: "Coding" was assigned to "word1", "is" was assigned to "word2", and "cool" was assigned to "word3". These three variables were then combined under one new variable called "phrase" using he method of sting concatenation (using the plus symbol to add the variables together). I also added quotation marks with nothing between them ("") so that the resulting phrase would have spaces between each two words. After that, I typed out a print function that called in the "phrase" variable. This should result in the output of the phrase Coding is cool"