Creating a program that outputs that creates a phrase using string concatenation and outputs it
I assigned each word to a variable: "Coding" was assigned to "word1", "is" was assigned "word2", "cool" was assigned "word3". I then combined these variables under a a new "phrase" variable using string concatenation (the + sign was used to add these words together. I also typed in "" to act as spaces in between each two words.) After that, I typed in a print function that called in the "phrase" variable. This should result in the output of the phrase "Coding is cool".