Answer for Exercise 2 by Jonathan Deuel C. Levardo
I first declared two variables. The first variable was named "num1" which was initialized with the integer "8". The second variable was named "num2" which was initialized with the integer "10". After declaring those two variables, I declared another variable: "sum". Under this "sum" variable, I added "num1" and "num2" together. After all that was done, I used the print function to print out the sum value to the console.