Answers for Exercise 4 by Jonathan Deuel C. Levardo
I started with creating a variable containing an input() function that asks the user for the name of the capital of a European country. Through the use of a string method (.lower()), the user's answer will always be stored as completely lowercase. This is to ensure that the user's correct answer will always match the true answer no matter the capitalization. I then created an If Else statement that will tell whether or not the user's answer is correct.

I copy pasted the code explained earlier and modified each of them to ask for the names of other European capitals. This entire submission asks for 10 different European capitals in total.

For fun, I created a "score" variable that is increased by 1 every time the user answers correctly. In the end, there is an If Else statement that outputs the user's score compared to the maximum possible score.