Answers for Exercise 5 by Jonathan Deuel C. Levardo
I first created a dictionary named "months" that contained key-value pairs of month numbers and how many days each month has.

After creating the dictionary, I used an input() function to ask the user to input a month number. The user's chosen number is stored in the variable "user_input". Next, I used another input() function to ask the user if it is a leap year or not. The user's response is stored in the variable "leap_year_check".

The user's responses are verified via an if else statement. I wrote this statement in a way that it will result three different responses: the number of days the user's chosen month has, a message stating that the user's number is invalid, and a message stating that the user's response to the leap year question is invalid.

The if statement checks if the user answered "yes" when asked if it is a leap year (the string method .lower() was used to ensure that the user's response will be verified correctly regardless of capitalization). If it is a leap year, the dictionary is updated so that key "'2'" will be paired with the new value, "29". After the update, another if else statement verifies if the user's number exists as a key in the dictionary. If "months.get(user_input)" does not result in "None", then the program prints out the value (the number of days) to the console. If "months.get(user_input)" does result in "None", then the message "Invalid Number!" is printed out to the console.

The elif statement is written the exact same way the if statement is except for a few key differences. The elif statement checks if the user answered "no" when asked if it is a leap year. This would not lead to the dictionary being updated. Instead, the code goes straight to verifying if whether or not the user's number is valid before printing out the results to the console.

The else statement takes care of the possibility that the user answered anything else other than "yes" or "no" when asked if it is a leap year. This would result in the message ""Invalid Response!" being printed out to the console.