Answers for Exercise 6 by Jonathan Deuel C. Levardo
I first created two variables. The first, "attempts", was assigned with the integer "5". The second, "password", was assigned with the string "12345".

Next, I created a while loop which is active as long as "attempts" is greater than 0. The first statement under this while loop is a new variable named "guess". This variable contains an input() function that asks the user to enter a 5-digit password. This leads to an if else statement that verifies if the user's response is correct. If the user's response ("guess") matches the true password ("password"), The message "Access Granted" is printed to the console and the loop is ended. If the user's response is not correct, the "attempts" variable is subtracted by one (to avoid any confusion, this is done through the else statement). If "attempts" has been reduced to zero, another if statement prints this message to the console: "You have exceeded your chances! The authorities have been notified."