Answers for Exercise 7 by Jonathan Deuel C. Levardo
For the first task, I wrote a basic for loop that has a range starting from "0" and ending before "51" with "digit" acting as the target variable. I used the print() function to print the target variable to the console which should result in each number in the range except for 51 being printed on a new line.

For the second task, I copy pasted the for loop from the first task and modified it so that it includes "-1" as the third argument for the range. i also switched the places of "50" and "0". "50" is now the first argument of the range and "0" is the second argument.

For the third task, I reused the same for loop except now the range only starts from "30" and ends before "51".

The for loop for the fourth task has a range that starts from "50" and ends before "8". "-2" acts as the third argument that states that the range would countdown by 2 when the target variable is printed.

I answered the fifth and final task with a for loop that has a range stating from "100" and ending before "205". The third argument is "5" which makes the range count up by 5 when the target variable is printed.