Answer for Exercise 8 by Jonathan Deuel C. Levardo
I first copy pasted the names and initialized them to a new list called "names".

For the variables, I created three. Variable "search" contains an input() that asks the user to enter a name. Variable "result" uses the in operator to verify if the name in "search" could be found in the list. Variable "position" contains a list index() method that searches the position of the name in "search".

All of this is applied through an if else statement. If the "result" variable turns out to be "True", an f-string is printed to the console. This f-string confirms that the user's entered name is in the list and gives out the position of the name in said list. If the "result" variable turns out to be "False", a different f-string is printed to the console which states that the user's entered name is not in the list.