Answer for Exercise 9 by Jonathan Deuel C. Levardo
Under the hello() function, I wrote a print() function that contains the greeting "Hello". As for the main() function, I wrote the hello() function under it. After filling the blanks under the previously defined functions, I copied "if __name__ == '__main__':" which would execute the main() function. This would in turn execute the hello() function. The final result is the message "Hello" being printed to the console.