Answer for Exercise 10 by Jonathan Deuel C. Levardo
First I defined the main() function. Under this function, I created a variable named "num" which contains an input() function where the user inputs their number as well as another soon to be defined function named "even_or_odd()". The even_or_odd() function contained the "num" variable as its parameter.

After writing out the main() function, I then defined the even_or_odd() function and once again included "num" as its parameter. Under this function, I wrote an if else statement that verifies if the value of "num" can be found in a range . This range counts up from 2 to 100000 by increments of 2. If the value of "num" can be found in this range, a message is printed to the console stating that the user's number is even. If the value of "num" cannot be found in this range, a different message is printed to the console stating that the user's number is odd.

Since the even_or_odd() function is under the main() function, all that is needed is to execute the main() function to run the whole code.